feat(SimpleGraph): Weaker condition for paths in acyclic graphs
`IsAcyclic.isPath_iff_chain'` defines a weaker condition for proving
that a walk is a path, in particular it shows that rather than all
vertices in the support of a walk being distinct, one must only show
that consecutive vertices are distinct. This leads to a simple
corollary that trails are paths in acyclic graphs.